# Client/Services/Swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PerfomanceProject.API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Methodology": {
      "get": {
        "tags": [
          "Methodology"
        ],
        "summary": "Get all Methodologies",
        "operationId": "Methodology_Get",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Methodology"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Methodology"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Methodology"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Methodology"
        ],
        "summary": "Create a New Methodology",
        "operationId": "Methodology_Post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Methodology"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Methodology"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Methodology"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Methodology"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Methodology"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Methodology"
                }
              }
            }
          }
        }
      }
    },
    "/api/Methodology/{id}": {
      "get": {
        "tags": [
          "Methodology"
        ],
        "summary": "Get Methodology by Id",
        "operationId": "Methodology_GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Methodology"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Methodology"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Methodology"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Methodology"
        ],
        "summary": "Modify a Methodology",
        "operationId": "Methodology_Put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Methodology"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Methodology"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Methodology"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Methodology"
        ],
        "summary": "Remove a Methodology",
        "operationId": "Methodology_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Positions": {
      "get": {
        "tags": [
          "Positions"
        ],
        "summary": "Get all Positions",
        "operationId": "Positions_Get",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Position"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Position"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Position"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Positions"
        ],
        "summary": "Create a New Position",
        "operationId": "Positions_Post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Position"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Position"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Position"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Position"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Position"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Position"
                }
              }
            }
          }
        }
      }
    },
    "/api/Positions/{id}": {
      "get": {
        "tags": [
          "Positions"
        ],
        "summary": "Get Position by Id",
        "operationId": "Positions_GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Position"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Position"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Position"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Positions"
        ],
        "summary": "Modify a Position",
        "operationId": "Positions_Put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Position"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Position"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Position"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Positions"
        ],
        "summary": "Remove a Position",
        "operationId": "Positions_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Users/{userId}/positions/{positionId}": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update a User's Position",
        "operationId": "Users_UpdatePosition",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          {
            "name": "positionId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Position"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Position"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Position"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Remove a User's Position",
        "operationId": "Users_RemovePosition",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          {
            "name": "positionId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Methodology": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ObjectId": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "machine": {
            "type": "integer",
            "format": "int32",
            "readOnly": true,
            "deprecated": true
          },
          "pid": {
            "type": "integer",
            "format": "int32",
            "readOnly": true,
            "deprecated": true
          },
          "increment": {
            "type": "integer",
            "format": "int32",
            "readOnly": true,
            "deprecated": true
          },
          "creationTime": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "Position": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}